split into files & roll over works kind of!

+++ main/tasks.py
@@ -11,6 +11,16 @@ import json
 FITBIT_ENDPOINTS = {
     'activities-heart-intraday':
     'https://api.fitbit.com/1/user/-/activities/heart/date/DATE/1d/1sec.json',
+    'activities-steps-intraday':
+    'https://api.fitbit.com/1/user/-/activities/steps/date/DATE/1d/1min.json',
+    'activities-calories-intraday':
+    'https://api.fitbit.com/1/user/-/activities/calories/date/DATE/1d/1min.json',
+    'activities-distance-intraday':
+    'https://api.fitbit.com/1/user/-/activities/distance/date/DATE/1d/1min.json',
+    'activities-floors-intraday':
+    'https://api.fitbit.com/1/user/-/activities/floors/date/DATE/1d/1min.json',
+    'activities-elevation-intraday':
+    'https://api.fitbit.com/1/user/-/activities/elevation/date/DATE/1d/1min.json'
 }
 
 
@@ -30,6 +40,7 @@ def update_fitbit(oh_id):
             response = requests.get(
                 'https://api.fitbit.com/1/user/-/profile.json',
                 headers=headers)
+            print(response.json())
             fitbit_data['user'] = response.json()['user']
             print('got fitbit userprofile')
         update_endpoints(oh_member, fitbit_data, headers, old_file_id)
@@ -41,7 +52,8 @@ def identify_start_data_fetch(fitbit_data, data_key):
             start_date = fitbit_data[data_key][-1]['date']
             del(fitbit_data[data_key][-1])
             return fitbit_data, start_date
-    return fitbit_data, fitbit_data['user']['memberSince']
+    #return fitbit_data, fitbit_data['user']['memberSince']
+    return fitbit_data, "2018-10-28" # THIS IS JUST FOR TESTING TO KEEP THE DATA MANAGEABLE!
 
 
 def get_single_endpoint(url, start_date, header):
@@ -54,6 +66,7 @@ def get_single_endpoint(url, start_date, header):
 
 
 def update_endpoints(oh_member, fitbit_data, header, old_file_id):
+    rollovers = 0
     try:
         for endpoint, url in FITBIT_ENDPOINTS.items():
             print('start processing {} endpoint'.format(endpoint))
@@ -72,7 +85,13 @@ def update_endpoints(oh_member, fitbit_data, header, old_file_id):
                     fitbit_data[endpoint].append(new_data)
                 else:
                     fitbit_data[endpoint] = [new_data]
-                start_date = start_date + datetime.timedelta(days=1)
+                next_date = start_date + datetime.timedelta(days=1)
+                if str(next_date)[:7] != str(start_date)[:7]:
+                    # rolling over to the next month!
+                    rollovers += 1
+                    break
+                else:
+                    start_date = next_date
             print('finished updating {}'.format(endpoint))
     except ValueError:
         print('encountered rate limit for fitbit!')
@@ -85,13 +104,42 @@ def update_endpoints(oh_member, fitbit_data, header, old_file_id):
             f.flush()
             f.seek(0)
             ohapi.api.upload_stream(
-                f, "fitbit-intraday.json", metadata={
+                f, "fitbit-intraday-{}.json".format(str(start_date)[:7]),
+                metadata={
                     "description": "Intraday Fitbit Data",
                     "tags": [
                                 "fitbit", 'intraday', 'heart rate',
                                 'Fitbit Intraday', 'activity']
                     }, access_token=oh_member.get_access_token())
-            ohapi.api.delete_file(
-                file_id=old_file_id,
-                access_token=oh_member.get_access_token())
+            if old_file_id:
+                ohapi.api.delete_file(
+                    file_id=old_file_id,
+                    access_token=oh_member.get_access_token())
         print('updated data for {}'.format(oh_member.oh_id))
+        if rollovers == 6:
+            # all endpoints rolled over, time to initialize a "new" file for
+            # next month!
+            print('rolled over for all endpoints, init new file')
+            new_fitbit_data = {
+                'user': fitbit_data['user']
+            }
+            new_month = str(start_date + datetime.timedelta(days=1))
+            for key in FITBIT_ENDPOINTS:
+                new_fitbit_data[key] = [{'date': str(new_month)}]
+            with tempfile.TemporaryFile() as f:
+                js = json.dumps(new_fitbit_data)
+                js = str.encode(js)
+                f.write(js)
+                f.flush()
+                f.seek(0)
+                ohapi.api.upload_stream(
+                    f, "fitbit-intraday-{}.json".format(str(new_month)[:7]),
+                    metadata={
+                        "description": "Intraday Fitbit Data",
+                        "tags": [
+                                    "fitbit", 'intraday', 'heart rate',
+                                    'Fitbit Intraday', 'activity']
+                        }, access_token=oh_member.get_access_token())
+            update_fitbit.apply_async(args=[oh_member.oh_id, ])
+            print('wrote empty file for next month')
+            print('enqueue next month!')
